Monadic Iota!
Iota is now extended to support a tuple in the state wherein the second
element of the tuple is a monadic value that uses monadic binds to
support changing the state.

Result: we can now thread state through the parser. And not just state -
any monad. A Writer monad could be used to record parser errors (HTML),
a Reader monad could be used to change the behavior based on input. In
the example parser, the CParser uses the Identity monad to count the
number of characters that have been emitted and inserts a comment after
emitting a 10th character.

Also, as a breaking change "embedInner" has been removed from the Iota
API. The semantics of it seemed unclear - is the just-parsed input to be
fed to the inner parser, emitted, or ignored? Previously it was fed to
the inner parser, but this behavior was clearly wrong (as in the case of
the quoted C string, it meant that the C parser could "see" the initial
delimiter of "[C|" and could act upon it.)

The version number should probably be higher than 0.5 now - with these
changes the parser is now very close to complete (at least for parsing
Text input, a String and ByteString parser would hopefully be trivial.)